# docs/swagger.yaml
basePath: /
definitions:
  ark_internal_http_dto.AIChatResponse:
    properties:
      reply:
        type: string
    type: object
...
  ark_internal_http_dto.RideAssistantResponse:
    properties:
      booking:
        $ref: '#/definitions/ark_internal_http_dto.RideAssistantBooking'
      reply:
        type: string
      session:
        $ref: '#/definitions/ark_internal_http_dto.RideAssistantSession'
      status:
        type: string
    type: object
...
  ark_internal_httpx.ErrorBody:
    properties:
      error:
        type: string
    type: object
...
  internal_http_handlers.aiChatReq:
    properties:
      message:
        type: string
      uid:
        type: string
    required:
    - message
    - uid
    type: object
...
  internal_http_handlers.createEventReq:
    properties:
      description:
        type: string
      from:
        description: RFC3339
        type: string
      title:
        type: string
      to:
        description: RFC3339
        type: string
    required:
    - from
    - title
    - to
    type: object
...
  internal_http_handlers.rideAssistantReq:
    properties:
      context_info:
        type: string
      message:
        type: string
      session_id:
        type: string
    required:
    - message
    type: object
...
host: localhost:8080
info:
  contact:
    name: ZooArk Team
  description: 'ZooArk 叫車平台後端 API。除了 POST /api/users、GET /health 和 GET /readyz 之外，所有
    endpoint 都需要 Firebase Auth ID Token（`Authorization: Bearer <token>`）。金額欄位以 TWD
    分（1/100 元）為單位。'
  title: ZooArk API
  version: 0.1.0
paths:
  /api/ai/chat:
    post:
      consumes:
      - application/json
      description: Sends a message to the Gemini-backed chat service on behalf of
        `uid`. Requests are rate-limited by the shared token budget; 429 is returned
        when the caller has exhausted their quota.
      parameters:
      - description: User id and message
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.aiChatReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.AIChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: AI chat (Gemini, token-metered)
      tags:
      - AI
  /api/assistant/ride/messages:
    post:
      consumes:
      - application/json
      description: Sends a user message to the ride-assistant state machine. The assistant
        responds with a reply, optionally a session snapshot for follow-up turns,
        and optionally a booking block once a ride has been committed.
      parameters:
      - description: Assistant message payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.rideAssistantReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.RideAssistantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Ride-assistant turn
      tags:
      - AI
  /api/calendar/events:
    post:
      consumes:
      - application/json
      description: Creates a calendar event for the authenticated user. `from`/`to`
        must be RFC3339 timestamps.
      parameters:
      - description: Event from/to/title/description
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.createEventReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ark_internal_http_dto.CalendarEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Create calendar event
      tags:
      - Calendar
  /api/calendar/events/{id}:
    delete:
      description: Removes the calendar event (and any ride schedule tied to it).
        Returns 204 on success.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Delete calendar event
      tags:
      - Calendar
    put:
      consumes:
      - application/json
      description: Replaces the stored calendar event with the supplied fields. Same
        validation rules as CreateEvent.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated event fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.editEventReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.CalendarEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Edit calendar event
      tags:
      - Calendar
  /api/calendar/schedules:
    get:
      description: Returns every calendar-backed ride schedule the authenticated user
        owns, flattened into the `schedules` array.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_http_handlers.schedulesListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: List schedules for current user
      tags:
      - Calendar
    post:
      consumes:
      - application/json
      description: Creates a scheduled ride order whose pickup window is derived from
        the referenced calendar event, and ties the two together so cancelling/deleting
        the event also untethers the order.
      parameters:
      - description: Event id + ride pickup/dropoff
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.createAndTieOrderReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ark_internal_http_dto.CalendarScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Tie a ride order to a calendar event
      tags:
      - Calendar
  /api/calendar/schedules/{event_id}/order:
    delete:
      description: Breaks the tie between a calendar event and its associated scheduled
        order. The event stays; the order is cancelled.
      parameters:
      - description: Calendar event ID
        in: path
        name: event_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Untie ride order from calendar event
      tags:
      - Calendar
  /api/driver/create:
    post:
      consumes:
      - application/json
      description: Creates the driver profile for the authenticated user (driver_id
        always comes from the Firebase token, never from the body). Starts the driver
        in status "available".
      parameters:
      - description: License info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_driver.createReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ark_internal_http_dto.DriverResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Create driver profile
      tags:
      - Driver
  /api/driver/status:
    patch:
      consumes:
      - application/json
      description: Changes the authenticated driver's status to one of available,
        on_trip, offline. Returns the new status echoed back.
      parameters:
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_driver.updateStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_driver.updateStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Update driver status
      tags:
      - Driver
  /api/me:
    delete:
      description: Hard-deletes the authenticated user's profile. There is no soft-delete
        today, so the row is removed immediately.
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Delete current user
      tags:
      - Auth & User
    get:
      description: Returns the profile of the user identified by the Firebase ID Token
        in the Authorization header.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Get current user
      tags:
      - Auth & User
    patch:
      consumes:
      - application/json
      description: Patches the display name of the currently authenticated user. Only
        the name field is editable today.
      parameters:
      - description: Updated name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.updateUserNameReq'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Update current user name
      tags:
      - Auth & User
  /api/notifications/register:
    post:
      consumes:
      - application/json
      description: Registers (or refreshes) an FCM device token for the authenticated
        user so push notifications can be delivered. Idempotent — re-registering the
        same token updates the row.
      parameters:
      - description: FCM token, platform, optional device_id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.ensureDeviceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_http_handlers.ensureDeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Register FCM device
      tags:
      - Notification
  /api/orders:
    post:
      consumes:
      - application/json
      description: Creates an on-demand (immediate) ride order for the authenticated
        passenger and enters the waiting queue.
      parameters:
      - description: Pickup/dropoff and ride type
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.createOrderReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Create instant order
      tags:
      - Order
  /api/orders/{id}/accept:
    post:
      description: Driver accepts a ride offer; transitions the order to `approaching`.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Accept order (driver)
      tags:
      - Order
  /api/orders/{id}/arrived:
    post:
      description: Driver signals arrival at the pickup location. Transitions the
        order from `approaching` to `arrived`.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Driver arrived at pickup
      tags:
      - Order
  /api/orders/{id}/cancel:
    post:
      description: Cancels the order as the passenger. For scheduled orders past the
        free-cancel deadline, `late_cancel:true` is returned so the client can warn
        the user that a penalty may apply.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Cancel order (passenger)
      tags:
      - Order
  /api/orders/{id}/claim:
    post:
      description: Driver claims a `scheduled` order; transitions it to `assigned`
        and binds the driver_id until the driver departs or cancels.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Claim scheduled order (driver)
      tags:
      - Order
  /api/orders/{id}/complete:
    post:
      description: Driver signals dropoff is complete; transitions the order from
        `driving` to `payment` so the passenger can settle the fare.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Trip complete (enter payment)
      tags:
      - Order
  /api/orders/{id}/deny:
    post:
      description: Driver declines the ride offer. The order is returned to the `waiting`
        queue for re-matching.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Deny order (driver)
      tags:
      - Order
  /api/orders/{id}/driver-cancel:
    post:
      consumes:
      - application/json
      description: Driver releases a previously-claimed `assigned` scheduled order
        back to `scheduled` so another driver can pick it up. Optional `reason` is
        recorded on the order history.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Optional reason string
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_http_handlers.driverCancelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Release claimed scheduled order (driver)
      tags:
      - Order
  /api/orders/{id}/match:
    post:
      description: Temporary MVP endpoint that atomically matches an order in `waiting`
        with the authenticated driver and transitions it to `approaching`. Will be
        replaced by the dispatcher.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Match order to driver (MVP)
      tags:
      - Order
  /api/orders/{id}/meet:
    post:
      description: Passenger has boarded; transitions the order from `arrived` to
        `driving` and the trip begins.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Passenger boarded
      tags:
      - Order
  /api/orders/{id}/pay:
    post:
      description: Temporary MVP endpoint that marks the order as paid and transitions
        it from `payment` to `complete`. Will be replaced by the real payment flow.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Pay for trip (MVP)
      tags:
      - Order
  /api/orders/{id}/status:
    get:
      description: Returns the current status, optimistic-lock version, and assigned
        driver_id (if any) for a single order.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.OrderStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Get order status
      tags:
      - Order
...
  /api/orders/scheduled/available:
    get:
      description: Returns scheduled orders within [from,to] that have not yet been
        claimed by a driver. Both `from` and `to` must be RFC3339 timestamps and `from`
        must be strictly before `to`.
      parameters:
      - description: RFC3339 window start (inclusive)
        in: query
        name: from
        required: true
        type: string
      - description: RFC3339 window end (exclusive)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_http_handlers.scheduledOrdersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: List claimable scheduled orders
      tags:
      - Order
...
schemes:
- http
- https
securityDefinitions:
  FirebaseAuth:
    description: Firebase ID Token, formatted as `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"